Convert Holder to tsx
